+++ src/html/applications.html
@@ -32,7 +32,7 @@
     ></script>
 
     <!-- IMPORT TEMPLATES -->
-    <script src="../../dist/ts/templates/common/templates.js" defer></script>
+    <script src="../../dist/ts/common/templates.js" defer></script>
     <script
       type="module"
       src="../../dist/ts/templates/FormChallenge/AdressInfoContainerPage.js"
