+++ src/html/form-validation.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" type="text/css" href="../css/index.css" />
     <link rel="stylesheet" type="text/css" href="../css/form-validation.css" />
 
-    <script src="../../dist/ts/templates/common/templates.js" defer></script>
+    <script src="../../dist/ts/common/templates.js" defer></script>
     <script type="module" src="../../dist/ts/form-validation.js" defer></script>
     <script
       type="module"
