+++ src/html/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" type="text/css" href="../css/index.css" />
     <script src="../js/index.js" defer></script>
-    <script src="../../dist/ts/templates/common/templates.js" defer></script>
+    <script src="../../dist/ts/common/templates.js" defer></script>
     <title>Learning Web</title>
   </head>
   <body>
