+++ src/ts/common/templates.ts
