+++ src/ts/final-form-challenge.ts
@@ -1,5 +1,11 @@
 // 🟦#03.CHALLENGE: CREATE A COMPLEX FORM WITH ALL FIELDS (NAME, EMAIL, PASSWORD, ADDRESS FORM, PAYMENT FORM, FILE INPUT, RADIO BUTTONS, CHECKBOXES ETC)
 
+import {
+  getRouteQueryParams,
+  injectQueryParams,
+  parseStringToBoolean,
+} from './common/utils.js';
+
 //#################################################
 //TYPES
 //#################################################
@@ -58,37 +64,6 @@ const paymentInfoContainerPage = document.getElementById(
 //#################################################
 //FUNCTIONS
 //#################################################
-
-const parseStringToBoolean = (str: string) => {
-  return str.toLowerCase() === 'true';
-};
-
-const getRouteQueryParams = (searchValue: string): string[] => {
-  const { search } = window.location;
-  const searchParsed = search.replace('?', '');
-  const searchedSplit = searchParsed.split('&');
-  const dialogParamPairKey = searchedSplit
-    .map((x) => {
-      const [key, value] = x.split('=');
-      if (key.includes(searchValue)) {
-        return [key, value];
-      }
-    })
-    .flat()
-    .filter((x) => x !== undefined);
-
-  return dialogParamPairKey;
-};
-
-const injectQueryParams = (queryKey: ParamQueryEnum, value: any) => {
-  const url = new URL(window.location.href);
-  const searchParams = url.searchParams;
-
-  searchParams.set(queryKey, value);
-
-  // Update the browser's URL without reloading
-  history.pushState({}, '', `${url.pathname}?${searchParams.toString()}`);
-};
 
 const openDialogFn = () => {
   dialog.open = true;
